Explicit table for storing FTS text
According to "EXPLAIN QUERY PLAN", some SQL queries make SQLite
use FTS index (??) for ordinary rowid SELECTs. I am not sure,
if that is bad thing (and if so, why), but having separate
table for use in those queries should not result in much harm —
at low level SQLite always uses separate tables for storing FTS data,
the only difference is if they are explicit or not.