Make use of contentless FTS table new default
The latest IPTV version no longer relies on snippet() to
highlight matches in description. As such, there is no longer
any justification for not using contentless FTS table for
case-insensitive searches support. Make formely "legacy" format
a new default and remove ability to generate the EPGX in the other
format.